+++ Assets/Testing/StateMachines/StateMachineTest.cs
@@ -7,6 +7,11 @@ public class StateMachineTest : MonoBehaviour
 {
     private SimpleStateMachine stateMachine;
 
+
+    [SerializeField]
+    private TestDataObject testDataObject = new TestDataObject();
+    private StateMachine<TestDataObject> classBasedStateMachine;
+
     private void Awake()
     {
         stateMachine = new SimpleStateMachine("StateXSin", StateXSin)
@@ -15,11 +20,32 @@ public class StateMachineTest : MonoBehaviour
             .AddTransition("StateXSin", "StateYSin", CycleStateTransitionTest)
             .AddTransition("StateYSin", "StateZSin", CycleStateTransitionTest)
             .AddTransition("StateZSin", "StateXSin", CycleStateTransitionTest);
+
+
+        classBasedStateMachine = new StateMachine<TestDataObject>(testDataObject)
+            .AddState("AState", new TestStateA()
+                .AddTransition("BState", (data) => { return data.testData == "B"; })
+                .AddTransition("CState", (data) => { return data.testData == "C"; })
+                .AddTransition("DState", (data) => { return data.testData == "D"; }))
+            .AddState("BState", new TestStateB()
+                .AddTransition("AState", (data) => { return data.testData == "A"; })
+                .AddTransition("CState", (data) => { return data.testData == "C"; })
+                .AddTransition("DState", (data) => { return data.testData == "D"; }))
+            .AddState("CState", new TestStateC()
+                .AddTransition("AState", (data) => { return data.testData == "A"; })
+                .AddTransition("BState", (data) => { return data.testData == "B"; })
+                .AddTransition("DState", (data) => { return data.testData == "D"; }))
+            .AddState("DState", new TestStateD()
+                .AddTransition("AState", (data) => { return data.testData == "A"; })
+                .AddTransition("BState", (data) => { return data.testData == "B"; })
+                .AddTransition("CState", (data) => { return data.testData == "C"; }));
+
     }
 
     private void Update()
     {
-        stateMachine.Run();
+        //stateMachine.Run();
+        classBasedStateMachine.Tick();
     }
 
     private bool CycleStateTransitionTest(SimpleStateMachine stateMachine) 
@@ -42,3 +68,46 @@ public class StateMachineTest : MonoBehaviour
         transform.position = new Vector3(transform.position.x, transform.position.y, Mathf.Sin(Time.timeSinceLevelLoad));
     }
 }
+
+[System.Serializable]
+public class TestDataObject
+{
+    public string testData = "A";
+}
+
+#region ClassStates
+
+public class TestStateA : StateMachine<TestDataObject>.State
+{
+    protected override void Tick()
+    {
+        Debug.Log("StateA");
+    }
+}
+
+public class TestStateB : StateMachine<TestDataObject>.State
+{
+    protected override void Tick()
+    {
+        Debug.Log("StateB");
+    }
+}
+
+public class TestStateC : StateMachine<TestDataObject>.State
+{
+    protected override void Tick()
+    {
+        Debug.Log("StateC");
+    }
+}
+
+public class TestStateD : StateMachine<TestDataObject>.State
+{
+    protected override void Tick()
+    {
+        Debug.Log("StateD");
+    }
+}
+
+
+#endregion