+++ Assets/BulkTools/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changes:
 
+## [0.1.3]
+
+### Added:
+- Added new class based state machine
+  - Uses objects instead of methods for states
+  - Enforces use of transitions
+
+
+
 ## [0.1.2]
 
 ### Added: